BE: Adding Update Contact to the entity class FE: minor changes
